Split JICF into Sub-Modules (#102)
Closes #96 

Divides the monolithic `jicf.py` file into dedicated modules with clearly defined responsibilities under the `models/jicf/` directory:

- `profile.py`: Analytical JICF profile calculations.
- `generate.py`: Utilities for generating a JICF model as a pre-processing step.
- `source.py`: Runtime calculations of injector source terms and tracer particle convection.
- `plot.py`: Utilities for analyzing and visually inspecting the generated model. (Future work)

Also started vectorizing and annotating some of the methods.
