+++ src/stanshock/models/jicf/__init__.py
@@ -1,0 +1,7 @@
+from __future__ import annotations
+
+__all__ = ["AnalyticJICF", "FuelInjector", "JICFChemistrySource", "JICModel"]
+
+from stanshock.models.jicf.generate import JICModel
+from stanshock.models.jicf.profile import AnalyticJICF
+from stanshock.models.jicf.source import FuelInjector, JICFChemistrySource
